Added class in HTML file and refactored CSS styles
Added a class for the p tag inside of the success
message container. This is to be able to continue
styling the success message. And the CSS was
refactored to remove redundant styles from certain
classes. Specifically the color property could be
removed for certain classes, since the same colors
are being applied to all the h1 and p tags in the
project. So the unnecessary color properties were
removed and they were instead added to the h1
and p tags, targeting all the necessary tags
needed for styling.