Added more CSS and JS and refactored HTML file
In the JS file we added code within the submit
event listener to work with the span that we
added in the success message p text in the HTML
file. This is so that the text of the email is
dynamic, being able to show the email that the
user entered, and not placeholder text.
In the CSS file styles were applied to the
span for the email, to match closer to the design.