Adjust error message and success message code
Error message was positioned to be more aligned
with the email input below it. The JavaScript file
was refactored to check for the window.innerWidth,
to deal with a bug where the desktop newsletter
image and the mobile newsletter image were
overlapping when the mobile version of the success
message was dismissed. The changes fixed this.